+++ inst/shiny_app/server.R
@@ -1,9 +1,11 @@
 source(file.path("server", "data_preview.R"), local = TRUE, chdir = TRUE)
+source(file.path("server", "plot_settings.R"), local = TRUE, chdir = TRUE)
 
 
 # Main server function ---------------------------------------------------------
 server <- function(input, output, session) {
     data_preview(input, output, session)
+    plot_settings(input, output, session)
 
 
     output[["plot"]] <- renderPlot({