+++ inst/shiny_app/ui/sidebar_panel_ui.R
@@ -1,6 +1,3 @@
-source("plot_settings_ui.R", local = TRUE)
-
-
 # Sidebar panel UI -------------------------------------------------------------
 sidebar_panel_ui <- function() {
     sidebarPanel(
@@ -20,5 +17,14 @@ input_settings_ui <- function() {
     tabPanel(
         "Input Settings",
         tags$p("This is Input Settings tab.")
+    )
+}
+
+
+# Plot Settings tab UI ---------------------------------------------------------
+plot_settings_ui <- function() {
+    tabPanel(
+        "Plot Settings",
+        uiOutput("plot_settings_inputs")
     )
 }